Added env manipulations
