+++ srcs/utils/free.c
@@ -1,29 +1,23 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   lexer.h                                            :+:      :+:    :+:   */
+/*   free.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: misargsy <misargsy@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/11/18 17:11:41 by knishiok          #+#    #+#             */
-/*   Updated: 2023/11/26 12:00:37 by misargsy         ###   ########.fr       */
+/*   Created: 2023/11/26 12:31:04 by misargsy          #+#    #+#             */
+/*   Updated: 2023/11/26 12:31:20 by misargsy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef LEXER_H
-# define LEXER_H
+#include "utils.h"
 
-# include "type.h"
-# include "error.h"
+void	free2darr(char **arr)
+{
+	size_t	i;
 
-void	destroy_token(t_token **token);
-bool	is_space(const char c);
-void	skip_spaces(const char **line);
-bool	is_metacharacter(const char c);
-bool	corresponds(const char *line, const char *metastr);
-
-void	tokenize(t_state *data, const char *line);
-
-void	print_tokens(t_token *token);
-
-#endif
+	i = 0;
+	while (arr[i] != NULL)
+		free(arr[i++]);
+	free(arr);
+}
